Move TheMakeGenericExMethod test to TypeExtensionsFacts
